refactor: cleanup implements
