+++ src/libs/common/dom/index.ts
@@ -1,3 +1,5 @@
 export * from './add-class';
 export * from './get-elements';
+export * from './register-element-management';
 export * from './remove-element';
+export * from './run-on-document-ready';
