+++ src/libs/common/dom/register-element-management.ts
