+++ src/libs/common/string/remove-head.ts
