+++ src/libs/common/validation/is-empty.spec.ts
