+++ src/libs/common/validation/is-empty.ts
