--- src/libs/scrapbox/config.ts
@@ -1,6 +1,0 @@
-/**
- * config for scrapbox.io UserScript
- */
-export const config = {
-  textareaSelector: '#text-input',
-};
