--- src/libs/scrapbox/selectors.ts
@@ -1,4 +1,0 @@
-// element selectors for DOM of scrapbox.io
-export const selectors = {
-  linesInPage: '.page .lines > .line',
-};
