+++ src/libs/scrapbox/text/tag.ts
@@ -5,3 +5,5 @@ export const makeTag = (text: string) => {
 
   return text[0] !== '#' ? `#${text}` : text;
 };
+
+export const isTagString = (val: unknown) => typeof val === 'string' && val[0] === '#';
