--- src/libs/scrapbox/util.ts
@@ -1,1 +1,0 @@
-export const isTagString = (val: unknown) => typeof val === 'string' && val[0] === '#';
