+++ src/scripts/debug-board/use-debug-board.ts
