--- src/scripts/enable-private-api/enable-private-api.ts
@@ -1,7 +1,0 @@
-import { getPrivateApi } from '../../libs/scrapbox';
-
-export const enablePrivateApi = async () => {
-  // debug
-  const api = await getPrivateApi();
-  (window as any).api = api;
-};
