--- src/scripts/enable-private-api/index.ts
@@ -1,1 +1,0 @@
-export { enablePrivateApi } from './enable-private-api';
