--- src/scripts/enable-structuralization/enable-structuralization.spec.ts
@@ -1,12 +1,0 @@
-import { setupBodyForTest } from '../../test-helpers';
-import { getPage } from './enable-structuralization';
-
-describe('getPage', () => {
-  beforeEach(() => {
-    setupBodyForTest();
-  });
-
-  it('should get page instance', () => {
-    const page = getPage();
-  });
-});
