+++ src/scripts/index.ts
@@ -1,7 +1,6 @@
-export * from './apply-login-css-class';
 export * from './component-manager';
 export * from './datetime-on-list';
-export * from './enable-private-api';
-export * from './tag-template-automation';
+export * from './debug-board';
 export * from './highlight-tags-in-list-item';
-export * from './use-debug-board';
+export * from './logged-in-css-class';
+export * from './tag-automation';
