+++ src/scripts/logged-in-css-class/apply-login-css-class.ts
