+++ src/scripts/tag-automation/open-dialog-and-write-tags.ts
